SIEM/Agent/requirements.txt
watchdog
requests
python-dotenv
vt-py
